ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sentence-transformers==3.1.1
torch==2.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
kafka-python==2.0.2
PyMuPDF==1.24.10
python-docx==1.1.2
spacy==3.7.6
numpy==1.26.4
redis==5.0.8
